POC crawler: Load crawlers from registered modules
Use the weather module to register an appropriate crawler that is now
loaded in the crawlers cmd application rather than using the
CrawlerFactory with predefined crawler classes.

With that, we are also using the scheduler class for both periodic and
immediated mode.
