+++ flirror/exceptions.py
@@ -4,12 +4,6 @@ class FlirrorConfigError(Exception):
     pass
 
 
-class CrawlerConfigError(Exception):
-    """Exception if a crawler could not be initialized correctly."""
-
-    pass
-
-
 class CrawlerDataError(Exception):
     """Exception if some data could not be retrieved by a crawler."""
